Fix validation crash, hardcoded startingChips, CSV bounty=0 bug; improve code quality
Bug fixes:
- Fix worker posting validation errors as results, causing UI crash (P1)
- Pass startingChips through BreakevenResult instead of hardcoding 10000 (P2)
- Fix CSV parser treating bounty=0 as falsy, overwriting with default (P2)

Code quality:
- Add vitest + happy-dom with 8 tests (worker validation, CSV parser)
- Replace tournament_type/payout_type strings with serde enums
- Extract magic numbers to named constants
- Widen MemoCache bitmask from u32 to u64
- Strengthen SortKey to literal union type
- Replace CalculateButton inline style with CSS class
- Clean up Monte Carlo comments
- Parallelize CI rust/frontend jobs, add vitest step

+++ icm-engine/src/icm_exact.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-type MemoCache = HashMap<(usize, u32), f64>;
+type MemoCache = HashMap<(usize, u64), f64>;
 
 /// Compute exact ICM equity using Malmuth-Harville model with bitmask memoization.
 /// Recursion depth is pruned by the number of payout positions.
@@ -25,7 +25,7 @@ fn compute_player_equity(
     cache: &mut MemoCache,
 ) -> f64 {
     let n = stacks.len();
-    let eliminated: u32 = 0;
+    let eliminated: u64 = 0;
     let mut equity = 0.0;
 
     for (position, &payout) in payouts.iter().enumerate().take(num_paid) {
@@ -40,7 +40,7 @@ fn finish_probability(
     player: usize,
     position: usize,
     stacks: &[f64],
-    eliminated: u32,
+    eliminated: u64,
     n: usize,
     num_paid: usize,
     cache: &mut MemoCache,
@@ -82,7 +82,7 @@ fn finish_probability(
     prob
 }
 
-fn first_place_probability(player: usize, stacks: &[f64], eliminated: u32, n: usize) -> f64 {
+fn first_place_probability(player: usize, stacks: &[f64], eliminated: u64, n: usize) -> f64 {
     if eliminated & (1 << player) != 0 {
         return 0.0;
     }
